Polished hover interactions and close targets.
Now, when cursor hovers over a task, instead of a black border around the task, the close target is the only change.  The appearance of the close target is that which indicates the task under the cursor will be activated if clicked.

The close target now more faithfully matches the background of the task it sits upon.  It has a deep gray if sitting upon an inactive task and a gray/orange if sitting upon an active task.