Checkpoint my work.
I've added a click handler to both form buttons.  If the add task button is clicked, the intent is to simply add the task.  If the track time button is clicked, the intent is to add the task and start it's timer.  The submit action—for when the user presses Enter—remains the same.  I've no idea whether this works yet or not.