Changing input buttons from <input> to <button>.
I need to use images for the input buttons because unicode support is
not standard enough across browsers (and probably OSs).  Chrome, in
particular, apparently has poor unicode support.

To use images, I need to use HTML buttons for the form submit action
rather than generic HTML input elements.

I'm also including a new images directory with an imagine of a '+'
symbol and an image of a clock symbol.