Use unique div IDs for timer divs.
Timer div IDs were all 'timer.' It's not strictly legal to have multiple divs in the DOM with the same ID. I was fortunate to have coded the JS
in such a way that it all worked out. I don't want to rely on this.

So, now all timer divs have IDs unique'ified by the timer's task ID. Each timer div now has an ID in the following format: '<TaskID>_timer'.